fix: reduce collision when converting non-braced string interpolation
When declaring a regex like `*.matches(r'^[A-Z0-9]+$'.rules())` sanitization failed to interpret this and created `matches('^[A-Z0-9]+$(') && !)` when followed by another condition. It basically thought that `$'.` is a non-braces string interpolation.

Not sure how good the new regex is to recognise and replace. But this and the other "braced string interpolation" regex might need more work.