fix: json key rewriter only checks mixins for getter
Instead of checking the main interface or any other super type the rewriter was just checking mixin. This only works if the class we inspect has the `JsonKey` annotation in a mixin.

Now it tries to get the getter from all possible sources and adds a replacement for the first one having a `JsonKey` annotation attached.